requirements.txt
pyqt
pyqt5-stubs
